feat/fix: update script engine architecture, memory design, and httpserver config
