+++ ai/agent/dataadmintools_join_optimization_test.go
@@ -16,14 +16,14 @@ func TestRightSortOptimization_Valid(t *testing.T) {
 			},
 		},
 		// Join on f1
-		rightKeyFields: []string{"f1"},
+		rightKeyFields:   []string{"f1"},
 		rightKeyFieldMap: map[string]string{"f1": "f1"},
 		// Sort by f2 DESC
 		rightSortFields: []string{"f2 DESC"},
 	}
-	
+
 	jp.checkRightSortOptimization()
-	
+
 	if !jp.isRightSortOptimized {
 		t.Error("Expected optimization to be ENABLED")
 	}
@@ -39,14 +39,14 @@ func TestRightSortOptimization_Valid_WithPrefix(t *testing.T) {
 			},
 		},
 		// Join on f1 (user alias "a")
-		rightKeyFields: []string{"a"},
+		rightKeyFields:   []string{"a"},
 		rightKeyFieldMap: map[string]string{"a": "f1"},
 		// Sort by "b.f2 DESC"
 		rightSortFields: []string{"b.f2 DESC"},
 	}
-	
+
 	jp.checkRightSortOptimization()
-	
+
 	if !jp.isRightSortOptimized {
 		t.Error("Expected optimization to be ENABLED with stripped prefix")
 	}
@@ -60,14 +60,14 @@ func TestRightSortOptimization_Invalid_Direction(t *testing.T) {
 				{FieldName: "f2", AscendingSortOrder: true}, // ASC in Index
 			},
 		},
-		rightKeyFields: []string{"f1"},
+		rightKeyFields:   []string{"f1"},
 		rightKeyFieldMap: map[string]string{"f1": "f1"},
 		// Sort by f2 DESC (Mismatch)
 		rightSortFields: []string{"f2 DESC"},
 	}
-	
+
 	jp.checkRightSortOptimization()
-	
+
 	if jp.isRightSortOptimized {
 		t.Error("Expected optimization to be DISABLED (Direction Mismatch)")
 	}
@@ -82,14 +82,14 @@ func TestRightSortOptimization_Invalid_Gap(t *testing.T) {
 				{FieldName: "f3", AscendingSortOrder: true},
 			},
 		},
-		rightKeyFields: []string{"f1"},
+		rightKeyFields:   []string{"f1"},
 		rightKeyFieldMap: map[string]string{"f1": "f1"},
 		// Sort by f3 (Skip f2)
 		rightSortFields: []string{"f3 ASC"},
 	}
-	
+
 	jp.checkRightSortOptimization()
-	
+
 	if jp.isRightSortOptimized {
 		t.Error("Expected optimization to be DISABLED (Field Gap)")
 	}
