+++ ai/agent/project_integration_flow_test.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"encoding/json"
 	"testing"
+
 	"github.com/stretchr/testify/assert"
 )
