+++ ai/constants.go
@@ -5,4 +5,7 @@ const (
 	DefaultModelOpenAI = "gpt-5.2"
 	DefaultModelGemini = "gemini-3-pro-preview"
 	DefaultModelOllama = "llama3"
+
+	// Default Application Configurations
+	DefaultScriptCategory = "general"
 )
